Shore up the error stuff so tests pass
