+++ Sources/Phoenix/ChannelError.swift
@@ -1,0 +1,4 @@
+public enum ChannelError: Error {
+    case invalidJoinReply(Channel.Reply)
+    case isClosed
+}
