+++ Sources/Phoenix/WebSocketError.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public enum WebSocketError: Error {
+    case unopened
+    case invalidURL(URL)
+    case invalidURLComponents(URLComponents)
+    case notOpen
+    case closed(URLSessionWebSocketTask.CloseCode, Data?)
+}
